Make it a custom UserControl
